nautilus-agent/requirements.txt
requests>=2.31.0
PyPDF2>=3.0.1
openpyxl>=3.1.2
python-dotenv>=1.0.0
openai>=1.0.0
pillow>=10.0.0
PyGithub>=2.1.1
